Reinclusão de arquivos no projeto após alteração csproj Diogo
