--- UniRitter.UniRitter2015/Models/CommentModel.cs.orig
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
-
-﻿namespace UniRitter.UniRitter2015.Models
-{
-    public class CommentModel
-    {
-
-        public Guid? id { get; set; }
-
-        [Required]
-        [MaxLength(1000)]
-        public string body { get; set; }
-
-        [Required]
-        [MaxLength(50)]
-        public string title { get; set; }
-
-        public Guid? author { get; set; }
-
-    }
-}
