+++ UniRitter.UniRitter2015/Models/PostModel.cs
@@ -1,5 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Web;
+
 namespace UniRitter.UniRitter2015.Models
 {
     public class PostModel
