--- UniRitter.UniRitter2015/Models/PostModel.cs.orig
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
-
-﻿namespace UniRitter.UniRitter2015.Models
-{
-    public class PostModel
-    {
-
-        public Guid? id { get; set; }
-
-        [Required]
-        [MaxLength(4000)]
-        public string body { get; set; }
-
-        [Required]
-        [MaxLength(100)]
-        public string title { get; set; }
-
-        public Guid? author { get; set; }
-
-        public string[] tags {get; set;}
-
-    }
-}
